update primary key from UUID to Integer

+++ src/main/java/com/kh/rupp_dev/boukryuniversity/controller/AttendanceSessionController.java
@@ -20,7 +20,7 @@ public class AttendanceSessionController {
     }
 
     @PatchMapping("/{id}/close")
-    public void close(@PathVariable Long id) {
+    public void close(@PathVariable Integer id) {
         service.closeSession(id);
     }
