+++ src/main/java/com/kh/rupp_dev/boukryuniversity/dto/response/AttendanceSessionResponse.java
@@ -3,7 +3,7 @@ package com.kh.rupp_dev.boukryuniversity.dto.response;
 import java.time.LocalDateTime;
 
 public record AttendanceSessionResponse(
-        Long sessionId,
+        Integer sessionId,
         String qrToken,
         LocalDateTime startTime,
         LocalDateTime endTime
